requirements.txt
#Requirements

Any Relational database.
Here I'm using mysql 5.7.18

Python 2.7

Django-admin 1.10 or later(Using 1.11)
